\name{grid2ip.p}
\docType{data}
\alias{grid2ip.p}
\title{p-values from gene GRID2IP}
\description{p-values of the single-nucleotide polymorphisms (SNPs) in gene GRID2IP (Van Assche et al. 2017).}
\usage{grid2ip.p}
\format{The data frame contains the following objects:
\tabular{lll}{
   \bold{grid2ip.p} \tab \code{numeric} \tab p-values \cr
}
}
\details{
   An example data set for combining p-values in a single gene.
}
\references{
   Van Assche E, Moons T, Cinar O, Viechtbauer W, Oldehinkel AJ, Van Leeuwen K, Verschueren K, Colpin H, Lambrechts D, Van den Noortgate W, et al. (2017). “Gene-based interaction analysis shows GABA ergic genes interacting with parenting in adolescent depressive symptoms.” \emph{Journal of Child Psychology and Psychiatry}, 58(12), 1301–1309.
}
\keyword{datasets}
