\name{print.combP}
\alias{print.combP}
\title{Print Method for 'combP' Objects}
\description{
   Print method for objects of class \code{"combP"}.
}
\usage{
\method{print}{combP}(x, \dots)
}
\arguments{
   \item{x}{an object of class \code{"combP"}.}
   \item{\dots}{other arguments.}
}
\details{
   The output shows the combined p-value.
}
\value{
   The function does not return an object.
}
\author{
   Ozan Cinar \email{ozancinar86@gmail.com} \cr
   Wolfgang Viechtbauer \email{wvb@wvbauer.com} \cr
}
\keyword{print}
