\name{print.poolr}
\alias{print.poolr}
\title{Print Method for 'poolr' Objects}
\description{
   Print method for objects of class \code{"poolr"}.
}
\usage{
\method{print}{poolr}(x, digits=3, \dots)
}
\arguments{
   \item{x}{an object of class \code{"poolr"}.}
   \item{digits}{integer.}
   \item{\dots}{other arguments.}
}
\details{
   The output shows the combined p-value.
}
\value{
   The function does not return an object.
}
\author{
   Ozan Cinar \email{ozancinar86@gmail.com} \cr
   Wolfgang Viechtbauer \email{wvb@wvbauer.com} \cr
}
\keyword{print}
