\name{grid2ip}
\docType{data}
\alias{grid2ip}
\title{LD matrix and p-values from gene GRID2IP}
\description{Linkage Disequilibrium (LD) matrix among the single-nucleotide polymorphisms (SNPs) in gene GRID2IP and p-values of these SNPs (Van Assche et al. 2017).}

\usage{
	data(grid2ip)
}

\format{The data frame contains the following objects:
\tabular{lll}{
	\bold{ld_mat}     	\tab \code{matrix}  	\tab LD matrix \cr
	\bold{p_vals}   	\tab \code{numeric} 	\tab p-values \cr
}
}
\details{
   An example data set for combining p-values in a single gene.
}
\references{
   Van Assche E, Moons T, Cinar O, Viechtbauer W, Oldehinkel AJ, Van Leeuwen K, Verschueren K, Colpin H, Lambrechts D, Van den Noortgate W, et al. (2017). “Gene-based interaction analysis shows GABA ergic genes interacting with parenting in adolescent depressive symptoms.” \emph{Journal of Child Psychology and Psychiatry}, 58(12), 1301–1309.
}
\keyword{datasets}
