\name{poolr-package}
\alias{poolr-package}
\alias{poolr}
\docType{package}
\title{Methods for Pooling P-Values from (Dependent) Tests}
\description{The poolr package is a collection of functions for pooling/combining the results (i.e., p-values) from (dependent) hypothesis tests. Included are Fisher's method, Stouffer's method, the inverse chi-squared method, the Bonferroni method, Tippett's method, and the binomial test. Each method can be adjusted based on an estimate of the number of effective tests or using empirically derived null distribution using pseudo replicates. For Fisher's, Stouffer's, and the inverse chi-squared method, direct generalizations based on multivariate theory are also available (leading to Brown's method, Strube's method, and the generalized inverse chi-squared method).}
\author{Ozan Cinar \email{ozancinar86@gmail.com} \cr
   Wolfgang Viechtbauer \email{wvb@wvbauer.com} \cr
}
\references{
   Bland, J. M., & Altman, D. G. (1995). \emph{Multiple significance tests: The Bonferroni method}. British Medical Journal, 310(6973), 170.

   Brown, MB (1975): \emph{400: A method for combining non-independent, one-sided tests of significance}. Biometrics, 987-992.

   Fisher, R. A. (1932). \emph{Statistical Methods for Research Workers} (4th ed.). Edinburgh: Oliver and Boyd.

   Mehta, C. R., & Patel, N. R. (1998). \emph{Exact inference for categorical data}. Encyclopedia of Biostatistics, 2, 1411-1422.

   Strube, M. J. (1985). \emph{Combining and comparing significance levels from nonindependent hypothesis tests}. Psychological Bulletin, 97(2), 334.

   Tippett, L. H. C. (1931). \emph{Methods of Statistics}. London: Williams Norgate.

   Wilkinson, B. (1951). \emph{A statistical consideration in psychological research}. Psychological Bulletin, 48(2), 156-158.
}
\keyword{package}
