\name{poolr-package}
\alias{poolr-package}
\alias{poolr}
\docType{package}
\title{Package for Pooling the Results from (Dependent) Tests}
\description{The poolr package is a collection of functions for pooling the results (usually p-values) of tests, which may possibly be dependent.
}
\author{Ozan Cinar \email{ozancinar86@gmail.com} \cr
   Wolfgang Viechtbauer \email{wvb@wvbauer.com} \cr
}
\references{
 Bland, J. M., & Altman, D. G. (1995). \emph{Multiple significance tests: The Bonferroni method}. British Medical Journal, 310(6973), 170. 
 
 Brown, MB (1975): \emph{400: A method for combining non-independent, one-sided tests of significance}. Biometrics, 987-992. 
 
 Fisher, R. A. (1932). \emph{Statistical methods for research workers} (4th ed.). Oliver and Boyd. 
 
 Mehta, C. R., & Patel, N. R. (1998). \emph{Exact inference for categorical data}. Encyclopedia of biostatistics, 2, 1411-1422. 
 
 Strube, M. J. (1985). \emph{Combining and comparing significance levels from nonindependent hypothesis tests}. Psychological bulletin, 97(2), 334. 
 
 Tippett, L. H. C. (1931). \emph{The Methods of Statistics}. The methods of statistics. 
 
 Wilkinson, B. (1951). \emph{A statistical consideration in psychological research}. Psychological Bulletin, 48(2), 156-158. 
}
\keyword{package}
